Started Decorator Design Pattern going to add template and change attribute search
