+++ Phase 0/CRC/Interface Adaptors/Controllers/RecipeBook.txt
@@ -9,3 +9,4 @@ Collaborators:
 * Command
 * User
 * UserManager
+* Console
