+++ Phase 0/CRC/Use Cases/OpenSesame.txt
@@ -1,0 +1,7 @@
+Class Name: OpenSesame
+
+Responsibilities:
+* Format information of a particular dish in a string
+
+Collaborators:
+* Dish
