+++ src/main/java/AttributeDecorator.java
@@ -1,0 +1,8 @@
+abstract class AttributeDecorator implements Dish{
+    public abstract boolean hasAttribute(AttributeDecorator attribute); {
+    }
+
+    public abstract PlainDish getPlainDish(); {
+    }
+
+}
